requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
pytest==6.2.5
pytest-django==4.4.0
flake8==3.9.2